Fix some errors in p3

+++ p3/setup_k3d_cluster.sh
@@ -7,9 +7,7 @@ RESET="\033[0m"
 CLUSTER_NAME="ci-cluster"
 
 create_cluster() {
-    if k3d cluster list | grep -q "${CLUSTER_NAME}" > /dev/null; then
-        echo -e "${GREEN} k3d cluster ${CLUSTER_NAME} exists.${RESET}"
-    else
+    if ! k3d cluster list | grep -q "${CLUSTER_NAME}"; then
         echo -e "${GREEN} Creating k3d cluster ${CLUSTER_NAME}...${RESET}"
         k3d cluster create ${CLUSTER_NAME}
         echo -e "${GREEN} Creating k3d namespaces...${RESET}"
@@ -24,22 +22,25 @@ wait_for_argocd_pods() {
     total_pods=$(kubectl get pods -n argocd --no-headers=true | wc -l)
 
     while [[ "$desired_ready_count" -ne "$total_pods" ]]; do
-        echo "[INFO][ARGOCD] Waiting for all pods to be ready..."
-        sleep 5
-
+        echo "Waiting for all pods to be ready..."
         desired_ready_count=$(kubectl get pods -n argocd --no-headers=true | awk '/Running/ && /1\/1/ {++count} END {print count}')
         total_pods=$(kubectl get pods -n argocd --no-headers=true | wc -l)
+        sleep 5
     done
 }
 
 setup_argo() {
-    echo -e "${GREEN} Installing argocd...${RESET}"
+    echo -e "${GREEN} Installing argocd pod...${RESET}"
     kubectl apply -n argocd -f https://raw.githubusercontent.com/argoproj/argo-cd/stable/manifests/install.yaml
 
-    echo -e "${GREEN} Downloading argocd cli...${RESET}"
-    curl -sSL -o argocd-linux-amd64 https://github.com/argoproj/argo-cd/releases/latest/download/argocd-linux-amd64
-    sudo install -m 555 argocd-linux-amd64 /usr/local/bin/argocd
-    rm argocd-linux-amd64
+    if ! command -v argocd &> /dev/null; then
+        echo -e "${GREEN} Downloading argocd cli...${RESET}"
+        curl -sSL -o argocd-linux-amd64 https://github.com/argoproj/argo-cd/releases/latest/download/argocd-linux-amd64
+        sudo install -m 555 argocd-linux-amd64 /usr/local/bin/argocd
+        rm argocd-linux-amd64
+    else
+        sleep 5
+    fi
 
     wait_for_argocd_pods
 
@@ -59,7 +60,6 @@ setup_argo() {
     ARGO_PSW=$(kubectl -n argocd get secret argocd-initial-admin-secret -o jsonpath="{.data.password}" | base64 -d)
     echo -e "${GREEN} argocd admin password: ${ARGO_PSW}${RESET}"
 
-    echo -e "${GREEN} Login to argocd server...${RESET}"
     argocd login localhost:8080 --insecure --username admin --password "$ARGO_PSW"
 }
 
@@ -69,30 +69,22 @@ create_argocd_app() {
     argocd app create will --repo https://github.com/Cpaluszek/cpalusze.git --path app --dest-server https://kubernetes.default.svc --dest-namespace dev --upsert
 
     argocd app get will --grpc-web
-
     argocd app set will --sync-policy automated
     argocd app set will --auto-prune --allow-empty --grpc-web
 
-    while true; do
-        output=$(argocd app get will --grpc-web)
-        if [[ $output == *"Service"*"Healthy"* && $output == *"Deployment"*"Healthy"* ]]; then
-            echo "Both service and deployment are healthy"
-            break
-        else
-            echo "Waiting for the app to become healthy..."
-            sleep 2
-        fi
+    while ! (argocd app get will --grpc-web | grep -q "Service.*Healthy" && argocd app get will --grpc-web | grep -q "Deployment.*Healthy"); do
+        echo "Waiting for the app to become healthy..."
+        sleep 2
     done
+    echo "Both service and deployment are healthy"
 
     if sudo lsof -i :8888 -sTCP:LISTEN -t | grep -q 'kubectl'; then
-        echo "Another port-forwarding process on port 8888 is already running. Killing it..."
         sudo pkill -f 'kubectl.*port-forward.*8888'
         sleep 5
     fi
 
     echo "Starting port forwarding for will-app..."
     kubectl port-forward svc/will-app -n dev 8888:8888 &
-    sudo netstat -tulpn | grep :8888
 
     while ! curl -s http://localhost:8888 > /dev/null; do
         echo "Waiting for port-forwarding for will-app to be ready..."
@@ -104,10 +96,8 @@ create_argocd_app() {
 
 print_infos() {
     argocd app get will
-
-    kubctl get pods -n dev
-
-    kubctl describe pod will -n dev
+    kubectl get pods -n dev
+    kubectl describe pod will -n dev
 }
 
 create_cluster
